requirements.txt
asgiref==3.5.2
boto3==1.24.22
botocore==1.27.22
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.0.12
click==8.1.3
colorama==0.4.4
cryptography==37.0.4
Flask==2.1.2
Flask-MySQL==1.5.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.4
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
MarkupSafe==2.1.1
phonenumberslite==8.12.51
Pillow==9.1.1
psycopg2==2.9.3
psycopg2-binary==2.9.3
pycparser==2.21
PyMySQL==1.0.2
python-dateutil==2.8.2
python-decouple==3.6
python-dotenv==0.20.0
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug==2.1.2
whitenoise==6.2.0
zipp==3.8.0
